refractor: retire high-priority structural debt in operator backend hotspots
